Move the helm and os packages to pkg (#2351)
moved interal/helm and internal/os packages to pkg dir and resorted imports on affected files
