+++ pkg/helm/helmcli.go
@@ -10,7 +10,8 @@ import (
 	"regexp"
 	"strings"
 
-	vzos "github.com/verrazzano/verrazzano/platform-operator/internal/os"
+	vzos "github.com/verrazzano/verrazzano/pkg/os"
+
 	"go.uber.org/zap"
 )
